+++ pro/ai/mcp/core/src/main/kotlin/org/http4k/ai/mcp/protocol/VersionedMcpEntity.kt
@@ -7,7 +7,11 @@ import se.ansman.kotshi.JsonSerializable
  * A description of an entity taking part in the MCP protocol - can be a client or a server.
  */
 @JsonSerializable
-data class VersionedMcpEntity(val name: McpEntity, val version: Version) {
-    constructor(name: String, version: String) : this(McpEntity.of(name), Version.of(version))
+data class VersionedMcpEntity(val name: McpEntity, val version: Version, val title: String? = null) {
+    constructor(name: String, version: String, title: String? = null) : this(
+        McpEntity.of(name),
+        Version.of(version),
+        title
+    )
 }