+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API
 changes with their rationale when appropriate:
 
+### v6.13.1.0
+- **http4k-***: Upgrade some dependency versions
+- **http4k-mcp-core**: Add new draft spec features to model
+
 ### v6.13.0.0
 - **http4k-***: Upgrade some dependency versions
 - **http4k-server-undertow**: [Fix #1365] Gracefully handle unrecognized/invalid HTTP verbs